Extract static bitmap helpers from RequestHandler
