+++ picasso/src/main/java/com/squareup/picasso3/AssetRequestHandler.java
@@ -25,6 +25,7 @@ import okio.Okio;
 import okio.Source;
 
 import static android.content.ContentResolver.SCHEME_FILE;
+import static com.squareup.picasso3.BitmapUtils.decodeStream;
 import static com.squareup.picasso3.Picasso.LoadedFrom.DISK;
 import static com.squareup.picasso3.Utils.checkNotNull;
