+++ picasso/src/main/java/com/squareup/picasso3/MediaStoreRequestHandler.java
@@ -32,6 +32,9 @@ import static android.provider.MediaStore.Images.Thumbnails.FULL_SCREEN_KIND;
 import static android.provider.MediaStore.Images.Thumbnails.MICRO_KIND;
 import static android.provider.MediaStore.Images.Thumbnails.MINI_KIND;
 import static android.provider.MediaStore.Video;
+import static com.squareup.picasso3.BitmapUtils.calculateInSampleSize;
+import static com.squareup.picasso3.BitmapUtils.createBitmapOptions;
+import static com.squareup.picasso3.BitmapUtils.decodeStream;
 import static com.squareup.picasso3.MediaStoreRequestHandler.PicassoKind.FULL;
 import static com.squareup.picasso3.MediaStoreRequestHandler.PicassoKind.MICRO;
 import static com.squareup.picasso3.MediaStoreRequestHandler.PicassoKind.MINI;
