+++ picasso/src/main/java/com/squareup/picasso3/ResourceDrawableRequestHandler.java
@@ -20,6 +20,7 @@ import android.graphics.drawable.Drawable;
 import android.support.annotation.NonNull;
 import android.support.v4.content.ContextCompat;
 
+import static com.squareup.picasso3.BitmapUtils.isXmlResource;
 import static com.squareup.picasso3.Picasso.LoadedFrom.DISK;
 
 public final class ResourceDrawableRequestHandler extends RequestHandler {
